fixed most exceptions

+++ Plugin/Patches/__System/__IO/File.cs
@@ -18,11 +18,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref bool __result, string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = !string.IsNullOrEmpty(path) && path.IndexOfAny(FileSystem.ExternalHelpers.InvalidPathChars) < 0 && FileSystem.FileExists(path);
             return false;
         }
@@ -33,11 +35,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string sourceFileName, string destFileName, bool overwrite)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.Copy(PathOrig.GetFullPath(sourceFileName), PathOrig.GetFullPath(destFileName), overwrite);
             return false;
         }
@@ -48,11 +52,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref string __result, string sourceFileName, string destFileName, bool overwrite)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             string fullPathInternal = PathOrig.GetFullPath(sourceFileName);
             string fullPathInternal2 = PathOrig.GetFullPath(destFileName);
             FileSystem.OsHelpers.Copy(fullPathInternal, fullPathInternal2, overwrite);
@@ -66,11 +72,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.DeleteFile(path);
             return false;
         }
@@ -81,11 +89,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref FileSecurity __result)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = new FileSecurity();
             __result.AddAccessRule(new FileSystemAccessRule("Everyone", FileSystemRights.FullControl, AccessControlType.Allow));
             return false;
@@ -97,11 +107,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix()
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             return false;
         }
     }
@@ -111,11 +123,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref FileAttributes __result, string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = FileSystem.OsHelpers.GetFileAttributes(path) ?? FileAttributes.Normal;
             return false;
         }
@@ -126,11 +140,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string path, FileAttributes fileAttributes)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.SetFileAttributes(path, fileAttributes);
             return false;
         }
@@ -141,11 +157,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref DateTime __result, string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = FileSystem.OsHelpers.FileCreationTime(path);
             return false;
         }
@@ -156,11 +174,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string sourceFileName, string destFileName)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             if (sourceFileName == null)
             {
                 throw new ArgumentNullException("sourceFileName");
@@ -204,11 +224,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string sourceFileName, string destinationFileName, string destinationBackupFileName/*, bool ignoreMetadataErrors*/)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             if (sourceFileName == null)
             {
                 throw new ArgumentNullException("sourceFileName");
@@ -282,11 +304,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
-        public static bool Prefix(ref DateTime __result, string path)
+        static bool Prefix(ref DateTime __result, string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = FileSystem.OsHelpers.FileAccessTime(path);
             return false;
         }
@@ -297,11 +321,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(ref DateTime __result, string path)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             __result = FileSystem.OsHelpers.FileWriteTime(path);
             return false;
         }
@@ -312,11 +338,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string path, DateTime creationTime)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.SetFileCreationTime(path, creationTime);
             return false;
         }
@@ -327,11 +355,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string path, DateTime lastAccessTime)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.SetFileAccessTime(path, lastAccessTime);
             return false;
         }
@@ -342,11 +372,13 @@ static class File
     {
         static Exception Cleanup(MethodBase original, Exception ex)
         {
-            return AuxilaryHelper.Cleanup_IgnoreException(original, ex);
+            return PatcherHelper.Cleanup_IgnoreException(original, ex);
         }
 
         static bool Prefix(string path, DateTime lastWriteTime)
         {
+            if (!PatcherHelper.CallFromPlugin())
+                return true;
             FileSystem.OsHelpers.SetFileWriteTime(path, lastWriteTime);
             return false;
         }
